Refactor: RestTemplate 요청 분리
