+++ src/main/java/com/nhnacademy/marketgg/client/aspect/AspectUtils.java
@@ -7,9 +7,19 @@ import lombok.NoArgsConstructor;
 import org.springframework.web.context.request.RequestContextHolder;
 import org.springframework.web.context.request.ServletRequestAttributes;
 
+/**
+ * AOP 편의 메서드 모음입니다.
+ *
+ * @author 윤동열
+ */
 @NoArgsConstructor(access = AccessLevel.PRIVATE)
-public class AspectUtils {
+public final class AspectUtils {
 
+    /**
+     * 요청 데이터를 반환 받습니다.
+     *
+     * @return 요청 정보
+     */
     public static HttpServletRequest getHttpRequest() {
         ServletRequestAttributes requestAttributes =
             Objects.requireNonNull(
