+++ src/main/java/com/nhnacademy/marketgg/client/repository/impl/CartAdapter.java
@@ -1,12 +1,6 @@
 package com.nhnacademy.marketgg.client.repository.impl;
 
-import com.fasterxml.jackson.core.JsonProcessingException;
-import com.fasterxml.jackson.databind.ObjectMapper;
 import com.nhnacademy.marketgg.client.dto.request.ProductToCartRequest;
-import com.nhnacademy.marketgg.client.dto.response.CartProductResponse;
-import com.nhnacademy.marketgg.client.dto.response.common.ResponseUtils;
-import com.nhnacademy.marketgg.client.exception.auth.UnAuthenticException;
-import com.nhnacademy.marketgg.client.exception.auth.UnAuthorizationException;
 import com.nhnacademy.marketgg.client.repository.CartRepository;
 import java.util.Collections;
 import java.util.List;
@@ -21,8 +15,9 @@ import org.springframework.web.client.RestTemplate;
 
 /**
  * 장바구니 데이터를 처리하는 구현체입니다.
- *
  * {@link com.nhnacademy.marketgg.client.repository.CartRepository}
+ *
+ * @author 윤동열
  */
 @Component
 @RequiredArgsConstructor
@@ -32,53 +27,35 @@ public class CartAdapter implements CartRepository {
 
     private final String gatewayIp;
     private final RestTemplate restTemplate;
-    private final ObjectMapper mapper;
 
     @Override
-    public void addToCart(final ProductToCartRequest productAddRequest)
-        throws JsonProcessingException, UnAuthenticException, UnAuthorizationException {
+    public ResponseEntity<String> addToCart(final ProductToCartRequest productAddRequest) {
+
         HttpEntity<ProductToCartRequest> httpEntity = new HttpEntity<>(productAddRequest, getJsonHeader());
 
-        ResponseEntity<String> response =
-            restTemplate.postForEntity(gatewayIp + CART, httpEntity, String.class);
-
-        ResponseUtils.checkError(response, mapper);
-
+        return restTemplate.postForEntity(gatewayIp + CART, httpEntity, String.class);
     }
 
     @Override
-    public List<CartProductResponse> retrieveCart()
-        throws JsonProcessingException, UnAuthenticException, UnAuthorizationException {
-
+    public ResponseEntity<String> retrieveCart() {
         HttpEntity<Void> httpEntity = new HttpEntity<>(getJsonHeader());
-        ResponseEntity<String> response =
-            restTemplate.exchange(gatewayIp + CART, HttpMethod.GET, httpEntity, String.class);
 
-        ResponseUtils.checkError(response, mapper);
-
-        return ResponseUtils.getListData(response.getBody(), mapper);
+        return restTemplate.exchange(gatewayIp + CART, HttpMethod.GET, httpEntity, String.class);
     }
 
     @Override
-    public void updateAmount(final ProductToCartRequest productUpdateRequest)
-        throws UnAuthenticException, UnAuthorizationException, JsonProcessingException {
+    public ResponseEntity<String> updateAmount(final ProductToCartRequest productUpdateRequest) {
 
         HttpEntity<ProductToCartRequest> httpEntity = new HttpEntity<>(productUpdateRequest, getJsonHeader());
-        ResponseEntity<String> response =
-            restTemplate.exchange(gatewayIp + CART, HttpMethod.PATCH, httpEntity, String.class);
-
-        ResponseUtils.checkError(response, mapper);
+        return restTemplate.exchange(gatewayIp + CART, HttpMethod.PATCH, httpEntity, String.class);
     }
 
     @Override
-    public void delete(final List<Long> products)
-        throws UnAuthenticException, UnAuthorizationException, JsonProcessingException {
+    public ResponseEntity<String> delete(final List<Long> products) {
 
         HttpEntity<List<Long>> httpEntity = new HttpEntity<>(products, getJsonHeader());
-        ResponseEntity<String> response =
-            restTemplate.exchange(gatewayIp + CART, HttpMethod.DELETE, httpEntity, String.class);
 
-        ResponseUtils.checkError(response, mapper);
+        return restTemplate.exchange(gatewayIp + CART, HttpMethod.DELETE, httpEntity, String.class);
     }
 
     private HttpHeaders getJsonHeader() {
