+++ src/main/java/com/nhnacademy/marketgg/client/service/impl/DefaultMemberService.java
@@ -8,7 +8,7 @@ import com.nhnacademy.marketgg.client.dto.response.EmailExistResponse;
 import com.nhnacademy.marketgg.client.dto.response.EmailUseResponse;
 import com.nhnacademy.marketgg.client.dto.response.MemberSignupResponse;
 import com.nhnacademy.marketgg.client.dto.response.MemberUpdateToAuthResponse;
-import com.nhnacademy.marketgg.client.repository.AuthRepository;
+import com.nhnacademy.marketgg.client.repository.MemberInfoRepository;
 import com.nhnacademy.marketgg.client.repository.MemberRepository;
 import com.nhnacademy.marketgg.client.service.MemberService;
 import java.time.LocalDateTime;
@@ -23,7 +23,7 @@ import org.springframework.stereotype.Service;
 public class DefaultMemberService implements MemberService {
 
     private final MemberRepository memberRepository;
-    private final AuthRepository authRepository;
+    private final MemberInfoRepository memberInfoRepository;
     private final PasswordEncoder passwordEncoder;
 
     @Override
@@ -43,14 +43,14 @@ public class DefaultMemberService implements MemberService {
 
     @Override
     public void update(final MemberUpdateToAuth memberUpdateToAuth, final String sessionId) {
-        MemberUpdateToAuthResponse updateData = authRepository.update(memberUpdateToAuth, sessionId);
+        MemberUpdateToAuthResponse updateData = memberInfoRepository.update(memberUpdateToAuth, sessionId);
         memberRepository.update(updateData, sessionId);
     }
 
     @Override
     public void withdraw(final String sessionId) {
         MemberWithdrawRequest withdrawRequest = new MemberWithdrawRequest(LocalDateTime.now());
-        authRepository.withdraw(withdrawRequest, sessionId);
+        memberInfoRepository.withdraw(withdrawRequest, sessionId);
         memberRepository.withdraw(withdrawRequest);
     }
 
@@ -59,7 +59,7 @@ public class DefaultMemberService implements MemberService {
         memberSignupRequest.encodePassword(passwordEncoder);
 
         MemberSignupResponse response =
-            authRepository.signup(memberSignupRequest.getSignupRequestToAuth());
+            memberInfoRepository.signup(memberSignupRequest.getSignupRequestToAuth());
 
         memberRepository.signup(memberSignupRequest
             .getSignupRequestToShopMember(response.getUuid(), response.getReferrerUuid()));
@@ -67,17 +67,17 @@ public class DefaultMemberService implements MemberService {
 
     @Override
     public EmailExistResponse checkEmail(final EmailRequest emailRequest) {
-        return authRepository.checkEmail(emailRequest);
+        return memberInfoRepository.checkEmail(emailRequest);
     }
 
     @Override
     public EmailUseResponse useEmail(final EmailRequest emailRequest) {
 
-        EmailUseResponse emailUseResponse = authRepository.useEmail(emailRequest);
+        EmailUseResponse emailUseResponse = memberInfoRepository.useEmail(emailRequest);
         Map<String, EmailUseResponse> useEmailResultMap = new HashMap<>();
         useEmailResultMap.put("isUseEmail", emailUseResponse);
-        return useEmailResultMap.get("isUeEmail");
 
+        return useEmailResultMap.get("isUeEmail");
     }
 
 }
